requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-chroma
python-dotenv